refactor(virtual machine service): moved lookup methods (cost, reegion, size) into separate service
now, IVirtualMachineService is only handling VM CRUD, while IVirtualMachineLookupService serves the
lookup with values
